+++ gfx/drivers/vulkan.c
@@ -4567,7 +4567,7 @@ static bool vulkan_frame(void *data, const void *frame,
          1);
 #endif
    vulkan_filter_chain_set_frame_time_delta(
-         (vulkan_filter_chain_t*)vk->filter_chain, video_driver_get_frame_time_delta_usec());
+         (vulkan_filter_chain_t*)vk->filter_chain, (uint32_t)video_driver_get_frame_time_delta_usec());
 
    vulkan_filter_chain_set_original_fps(
          (vulkan_filter_chain_t*)vk->filter_chain, video_driver_get_original_fps());
